refactor: address code review feedback - split components, add useGearPurchase hook, use form metadata
Agent-Logs-Url: https://github.com/CptnFizzbin/shadow-sin/sessions/5d9c03d8-cb9f-4448-a8dd-99907986be32
