+++ src/components/characterBuilder/sections/gear/cyberware/forms/useImplantForm.tsx
@@ -2,6 +2,11 @@ import type { UUID } from "node:crypto"
 
 import { createFieldMap, formOptions } from "@tanstack/form-core"
 
+import type { GearSubmitMeta } from "#/components/gear/gearSubmitMeta.ts"
+import {
+  defaultGearSubmitMeta,
+
+} from "#/components/gear/gearSubmitMeta.ts"
 import { useAppForm } from "#/integrations/tanstackForm/useAppForm.ts"
 import type { ImplantData } from "#/lib/system/gear/implantData.ts"
 import { ImplantGrade, ImplantLocation, ImplantType } from "#/lib/system/gear/implantData.ts"
@@ -11,7 +16,7 @@ import { NullUuid } from "#/lib/uuidUtils.ts"
 export interface ImplantFormOptions {
   implant?: ImplantData
   parentId?: UUID
-  onSubmit: (implant: ImplantData) => void
+  onSubmit: (implant: ImplantData, meta: GearSubmitMeta) => void
 }
 
 const defaultFormValues: ImplantData = {
@@ -63,6 +68,7 @@ export const useImplantForm = (options: ImplantFormOptions) => {
       parentId: options.parentId,
       ...options.implant,
     },
-    onSubmit: ({ value }) => options.onSubmit(value),
+    onSubmitMeta: defaultGearSubmitMeta,
+    onSubmit: ({ value, meta }) => options.onSubmit(value, meta),
   })
 }
