+++ src/components/gear/gearSubmitMeta.ts
@@ -1,0 +1,7 @@
+export type GearSubmitAction = "acquire" | "purchase" | "save"
+
+export type GearSubmitMeta = {
+  submitAction: GearSubmitAction
+}
+
+export const defaultGearSubmitMeta: GearSubmitMeta = { submitAction: "save" }
