+++ src/components/licenses/dialogs/licenseFormDialog.tsx
@@ -7,7 +7,6 @@ import DialogContent from "@mui/material/DialogContent"
 import DialogTitle from "@mui/material/DialogTitle"
 import Stack from "@mui/material/Stack"
 import type { FC } from "react"
-import { useRef } from "react"
 
 import { AvailabilityChip } from "#/components/gear/availabilityChip.tsx"
 import { GearAcquireActions } from "#/components/gear/gearAcquireActions.tsx"
@@ -44,15 +43,15 @@ export const LicenseFormDialog: FC<LicenseFormDialogProps> = ({
   sin,
 }) => {
   const title = license ? "Edit License" : "Create License"
-  const submitModeRef = useRef<"acquire" | "purchase" | "save">("save")
+  const useAcquireMode = !!onAcquire && !!onPurchase
 
   const form = useLicenseForm({
     license: license,
     parentId: sin?.id,
     sinReal: sin?.rating === "real" || false,
-    onSubmit: (submittedLicense) => {
-      if (onAcquire && onPurchase) {
-        if (submitModeRef.current === "purchase") {
+    onSubmit: (submittedLicense, meta) => {
+      if (useAcquireMode) {
+        if (meta.submitAction === "purchase") {
           onPurchase(submittedLicense)
         } else {
           onAcquire(submittedLicense)
@@ -62,8 +61,6 @@ export const LicenseFormDialog: FC<LicenseFormDialogProps> = ({
       }
     },
   })
-
-  const useAcquireMode = !!onAcquire && !!onPurchase
 
   return (
     <Dialog open={open} fullWidth onTransitionExited={onClosed}>
@@ -111,14 +108,8 @@ export const LicenseFormDialog: FC<LicenseFormDialogProps> = ({
                     <GearAcquireActions
                       cost={cost}
                       onClose={onClose}
-                      onAcquire={() => {
-                        submitModeRef.current = "acquire"
-                        form.handleSubmit()
-                      }}
-                      onPurchase={() => {
-                        submitModeRef.current = "purchase"
-                        form.handleSubmit()
-                      }}
+                      onAcquire={() => form.handleSubmit({ submitAction: "acquire" })}
+                      onPurchase={() => form.handleSubmit({ submitAction: "purchase" })}
                     />
                   )
                 }}
@@ -132,10 +123,7 @@ export const LicenseFormDialog: FC<LicenseFormDialogProps> = ({
                 <Button
                   type="submit"
                   color="secondary"
-                  onClick={() => {
-                    submitModeRef.current = "save"
-                    form.handleSubmit()
-                  }}
+                  onClick={() => form.handleSubmit()}
                   variant="contained"
                 >
                   Save
