+++ src/components/licenses/dialogs/sinFormDialog.tsx
@@ -7,7 +7,6 @@ import DialogContent from "@mui/material/DialogContent"
 import DialogTitle from "@mui/material/DialogTitle"
 import Stack from "@mui/material/Stack"
 import type { FC } from "react"
-import { useRef } from "react"
 
 import { AvailabilityChip } from "#/components/gear/availabilityChip.tsx"
 import { GearAcquireActions } from "#/components/gear/gearAcquireActions.tsx"
@@ -40,13 +39,13 @@ export const SinFormDialog: FC<SinFormDialogProps> = ({
   onDelete,
 }) => {
   const title = sin ? "Edit SIN" : "Create SIN"
-  const submitModeRef = useRef<"acquire" | "purchase" | "save">("save")
+  const useAcquireMode = !!onAcquire && !!onPurchase
 
   const form = useSinForm({
     sin,
-    onSubmit: (submittedSin) => {
-      if (onAcquire && onPurchase) {
-        if (submitModeRef.current === "purchase") {
+    onSubmit: (submittedSin, meta) => {
+      if (useAcquireMode) {
+        if (meta.submitAction === "purchase") {
           onPurchase(submittedSin)
         } else {
           onAcquire(submittedSin)
@@ -56,8 +55,6 @@ export const SinFormDialog: FC<SinFormDialogProps> = ({
       }
     },
   })
-
-  const useAcquireMode = !!onAcquire && !!onPurchase
 
   return (
     <Dialog open={open} fullWidth onTransitionExited={onClosed}>
@@ -110,14 +107,8 @@ export const SinFormDialog: FC<SinFormDialogProps> = ({
                     <GearAcquireActions
                       cost={cost}
                       onClose={onClose}
-                      onAcquire={() => {
-                        submitModeRef.current = "acquire"
-                        form.handleSubmit()
-                      }}
-                      onPurchase={() => {
-                        submitModeRef.current = "purchase"
-                        form.handleSubmit()
-                      }}
+                      onAcquire={() => form.handleSubmit({ submitAction: "acquire" })}
+                      onPurchase={() => form.handleSubmit({ submitAction: "purchase" })}
                     />
                   )
                 }}
@@ -131,10 +122,7 @@ export const SinFormDialog: FC<SinFormDialogProps> = ({
                 <Button
                   type="submit"
                   color="secondary"
-                  onClick={() => {
-                    submitModeRef.current = "save"
-                    form.handleSubmit()
-                  }}
+                  onClick={() => form.handleSubmit()}
                   variant="contained"
                 >
                   Save
